requirements.txt
python-dotenv
yt-dlp
aiogram
requests
moviepy
instaloader
psycopg2-binary
httpx